* If you resurrect in Red River Rafting while riding a raft, you go to the beginning and both logs reset position * If you're riding a log, the camera looks downward no matter which way you face * replaced the 'water' image in red river rafting with another one I found in the old kid mystic folder * Rune added 4/6 * major tweaks to how water works. 50/50 on whether this is going to work well

+++ source/mystic/water.h
@@ -8,7 +8,7 @@ void ExitWater(void);
 void UpdateWater(void);
 byte WaterPixel(int x,int y);
 void WaterRipple(int x,int y,short amt);
-void WaterBlop(byte x,byte y,byte width);
+void WaterBlop(int x,int y,int width);
 void SetupWater(void);
 
 #endif
